requirements.txt
streamlit==1.31.0
google-generative-ai==0.7.2
python-dotenv==1.0.1